chore: moved to index.js
